refactor: add projects component
