+++ src/components/Projects/index.ts
@@ -1,0 +1,3 @@
+import ProjectsContainer from './ProjectsContainer';
+
+export default ProjectsContainer;
